Fix signed/unsigned comparison warning.

+++ lexers/LexHTML.cxx
@@ -2134,7 +2134,7 @@ static void ColouriseHyperTextDoc(unsigned int startPos, int length, int initSty
 		break;
 	default:
 		StateToPrint = statePrintForState(state, inScriptType);
-		if (styler.GetStartSegment() < lengthDoc)
+		if (static_cast<int>(styler.GetStartSegment()) < lengthDoc)
 			styler.ColourTo(lengthDoc - 1, StateToPrint);
 		break;
 	}
